+++ R/write.R
@@ -27,7 +27,7 @@
 #' Three extensions are currently supported: `.gz` for gzip compression, `.bz2` for bzip2 compression and
 #' `.xz` for lzma compression.  See the examples for more information.
 #'
-#' @param x A data frame or tibble to write to disk
+#' @param x A data frame or tibble to write to disk.
 #' @param path The path or connection to write to.
 #' @param append If `FALSE`, will overwrite existing file. If `TRUE`,
 #'   will append to existing file. In both cases, if the file does not exist a new
@@ -42,6 +42,8 @@
 #' @param quote_escape The type of escaping to use for quoted values, one of
 #'   `"double"`, `"backslash"` or `"none"`. You can also use `FALSE`, which is
 #'   equivalent to "none". The default is `"double"`, which is expected format for Excel.
+#' @param eol The end of line character to use. Most commonly either `"\n"` for
+#'   Unix style newlines, or `"\r\n"` for Windows style newlines.
 #' @return `write_*()` returns the input `x` invisibly.
 #' @references Florian Loitsch, Printing Floating-Point Numbers Quickly and
 #' Accurately with Integers, PLDI '10,
@@ -62,13 +64,13 @@
 #'
 #' \dontshow{setwd(.old_wd)}
 write_delim <- function(x, path, delim = " ", na = "NA", append = FALSE,
-                        col_names = !append, quote_escape = "double") {
+                        col_names = !append, quote_escape = "double", eol = "\n") {
   stopifnot(is.data.frame(x))
 
   x_out <- x
   x[] <- lapply(names(x), function(i) output_column(x[[i]], i))
   stream_delim(x, path, delim = delim, col_names = col_names, append = append,
-    na = na, quote_escape = quote_escape)
+    na = na, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
@@ -76,19 +78,19 @@ write_delim <- function(x, path, delim = " ", na = "NA", append = FALSE,
 #' @rdname write_delim
 #' @export
 write_csv <- function(x, path, na = "NA", append = FALSE, col_names = !append,
-                      quote_escape = "double") {
+                      quote_escape = "double", eol = "\n") {
   write_delim(x, path, delim = ",", na = na, append = append,
-    col_names = col_names, quote_escape = quote_escape)
+    col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' @rdname write_delim
 #' @export
 write_csv2 <- function(x, path, na = "NA", append = FALSE, col_names = !append,
-                       quote_escape = "double") {
+                       quote_escape = "double", eol = "\n") {
   x_out <- x
   x <- change_decimal_separator(x, decimal_mark = ",")
   write_delim(x, path, delim = ";", na = na, append = append,
-    col_names = col_names, quote_escape = quote_escape)
+    col_names = col_names, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
@@ -96,7 +98,7 @@ write_csv2 <- function(x, path, na = "NA", append = FALSE, col_names = !append,
 #' @rdname write_delim
 #' @export
 write_excel_csv <- function(x, path, na = "NA", append = FALSE,
-                            col_names = !append, delim = ",", quote_escape = "double") {
+                            col_names = !append, delim = ",", quote_escape = "double", eol = "\n") {
 
   stopifnot(is.data.frame(x))
 
@@ -106,7 +108,7 @@ write_excel_csv <- function(x, path, na = "NA", append = FALSE,
 
   x[] <- lapply(x, output_column)
   stream_delim(x, path, delim, col_names = col_names, append = append,
-    na = na, bom = TRUE, quote_escape = quote_escape)
+    na = na, bom = !append, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
@@ -114,7 +116,7 @@ write_excel_csv <- function(x, path, na = "NA", append = FALSE,
 #' @rdname write_delim
 #' @export
 write_excel_csv2 <- function(x, path, na = "NA", append = FALSE,
-                             col_names = !append, delim = ";", quote_escape = "double") {
+                             col_names = !append, delim = ";", quote_escape = "double", eol = "\n") {
   x_out <- x
   x <- change_decimal_separator(x, decimal_mark = ",")
 
@@ -122,15 +124,15 @@ write_excel_csv2 <- function(x, path, na = "NA", append = FALSE,
   x[datetime_cols] <- lapply(x[datetime_cols], format, "%Y/%m/%d %H:%M:%S")
 
   x[] <- lapply(x, output_column)
-  write_excel_csv(x, path, na, append, col_names, delim, quote_escape = quote_escape)
+  write_excel_csv(x, path, na, append, col_names, delim, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
 
 #' @rdname write_delim
 #' @export
-write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
-  write_delim(x, path, delim = '\t', na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
+  write_delim(x, path, delim = '\t', na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' Convert a data frame to a delimited string
@@ -141,6 +143,7 @@ write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, q
 #' @return A string.
 #' @inheritSection write_delim Output
 #' @inheritParams write_delim
+#' @param x A data frame.
 #' @inherit write_delim references
 #' @examples
 #' # format_()* functions are useful for testing and reprexes
@@ -157,32 +160,32 @@ write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, q
 #' cat(format_csv(df))
 #' @export
 format_delim <- function(x, delim, na = "NA", append = FALSE,
-                         col_names = !append, quote_escape = "double") {
+                         col_names = !append, quote_escape = "double", eol = "\n") {
   stopifnot(is.data.frame(x))
 
   x[] <- lapply(x, output_column)
-  res <- stream_delim(df = x, path = NULL, delim = delim, col_names = col_names, append = append, na = na, quote_escape = quote_escape)
+  res <- stream_delim(df = x, path = NULL, delim = delim, col_names = col_names, append = append, na = na, quote_escape = quote_escape, eol = eol)
   Encoding(res) <- "UTF-8"
   res
 }
 
 #' @export
 #' @rdname format_delim
-format_csv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
-  format_delim(x, delim = ",", na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+format_csv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
+  format_delim(x, delim = ",", na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' @export
 #' @rdname format_delim
-format_csv2 <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
+format_csv2 <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
   x <- change_decimal_separator(x, decimal_mark = ",")
-  format_delim(x, delim = ";", na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+  format_delim(x, delim = ";", na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' @export
 #' @rdname format_delim
-format_tsv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
-  format_delim(x, delim = "\t", na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+format_tsv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
+  format_delim(x, delim = "\t", na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' Preprocess column for output
@@ -222,7 +225,7 @@ output_column.list <- function(x, name) {
   stop("Flat files can't store the list column `", name, "`", call. = FALSE)
 }
 
-stream_delim <- function(df, path, append = FALSE, bom = FALSE, ..., quote_escape) {
+stream_delim <- function(df, path, append = FALSE, bom = FALSE, ..., quote_escape, eol) {
   quote_escape <- standardise_escape(quote_escape)
 
   path <- standardise_path(path, input = FALSE)
@@ -235,7 +238,7 @@ stream_delim <- function(df, path, append = FALSE, bom = FALSE, ..., quote_escap
       open(path, "wb")
     }
   }
-  stream_delim_(df, path, ..., bom = bom, quote_escape = quote_escape)
+  stream_delim_(df, path, ..., bom = bom, quote_escape = quote_escape, eol = eol)
 }
 
 change_decimal_separator <- function(x, decimal_mark = ",") {