+++ src/DateTimeParser.h
@@ -4,13 +4,15 @@
 #include "DateTime.h"
 #include "LocaleInfo.h"
 #include "QiParsers.h"
+#include "cpp11/protect.hpp"
+
 #include "boost.h"
 #include <ctime>
 
 // Parsing ---------------------------------------------------------------------
 
 class DateTimeParser {
-  int year_, mon_, day_, hour_, min_, sec_;
+  int sign_, year_, mon_, day_, hour_, min_, sec_;
   double psec_;
   int amPm_;
   bool compactDate_; // used for guessing
@@ -144,7 +146,7 @@ public:
       }
 
       if (formatItr + 1 == formatEnd)
-        Rcpp::stop("Invalid format: trailing %");
+        cpp11::stop("Invalid format: trailing %%");
       formatItr++;
 
       switch (*formatItr) {
@@ -181,9 +183,16 @@ public:
         if (!consumeInteger1WithSpace(2, &day_))
           return false;
         break;
-      case 'H': // hour
+      case 'h': // hour, unrestricted
+        if (!consumeHours(&hour_, &sign_))
+          return false;
+        break;
+      case 'H': // hour, 0-23
         if (!consumeInteger(2, &hour_, false))
           return false;
+        if (hour_ < 0 || hour_ > 23) {
+          return false;
+        }
         break;
       case 'I': // hour
         if (!consumeInteger(2, &hour_, false))
@@ -203,7 +212,7 @@ public:
         break;
       case 'O': // seconds (double)
         if (formatItr + 1 == formatEnd || *(formatItr + 1) != 'S')
-          Rcpp::stop("Invalid format: %%O must be followed by %%S");
+          cpp11::stop("Invalid format: %%O must be followed by %%S");
         formatItr++;
         if (!consumeSeconds(&sec_, &psec_))
           return false;
@@ -241,7 +250,7 @@ public:
 
       case 'A': // auto date / time
         if (formatItr + 1 == formatEnd)
-          Rcpp::stop("Invalid format: %%A must be followed by another letter");
+          cpp11::stop("Invalid format: %%A must be followed by another letter");
         formatItr++;
         switch (*formatItr) {
         case 'D':
@@ -253,7 +262,7 @@ public:
             return false;
           break;
         default:
-          Rcpp::stop("Invalid %%A auto parser");
+          cpp11::stop("Invalid %%A auto parser");
         }
         break;
 
@@ -276,7 +285,7 @@ public:
         break;
 
       default:
-        Rcpp::stop("Unsupported format %%%s", *formatItr);
+        cpp11::stop("Unsupported format %%%s", *formatItr);
       }
     }
 
@@ -297,7 +306,15 @@ public:
     return dt;
   }
   DateTime makeTime() {
-    DateTime dt(0, 0, 0, hour(), min_, sec_, psec_, "UTC");
+    DateTime dt(
+        0,
+        0,
+        0,
+        sign_ * hour(),
+        sign_ * min_,
+        sign_ * sec_,
+        sign_ * psec_,
+        "UTC");
     return dt;
   }
 
@@ -325,6 +342,26 @@ private:
 
     // 24 hour time
     return hour_;
+  }
+
+  inline bool consumeHours(int* pHour, int* pSign) {
+    if (dateItr_ == dateEnd_)
+      return false;
+
+    int sign = 1;
+
+    if (*dateItr_ == '-') {
+      sign = -1;
+      ++dateItr_;
+    } else if (*dateItr_ == '+') {
+      ++dateItr_;
+    }
+
+    if (!consumeInteger(10, pHour, false))
+      return false;
+
+    *pSign = sign;
+    return true;
   }
 
   inline bool consumeSeconds(int* pSec, double* pPartialSec) {
@@ -487,6 +524,7 @@ private:
   }
 
   void reset() {
+    sign_ = 1;
     year_ = -1;
     mon_ = 0;
     day_ = 0;