+++ tests/testthat/test-parsing-factors.R
@@ -53,7 +53,7 @@ test_that("NAs included in levels if desired", {
   expect_equal(x, factor(c("NA", "b", "a"), levels = c("b", "a")))
 
   x <- parse_factor(c("NA", "b", "a"), levels = NULL, include_na = TRUE)
-  expect_equal(x, factor(c("NA", "b", "a"), levels = c(NA, "b", "a"), exclude = NULL))
+  expect_equal(x, factor(c(NA, "b", "a"), levels = c(NA, "b", "a"), exclude = NULL))
 })
 
 test_that("Factors handle encodings properly (#615)", {